pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "office-powerpoint-mcp-server-s3"
version = "2.1.0"
description = "A MCP server for PowerPoint manipulation with S3 support"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    { name = "Athar Paymode", email = "your.email@example.com" }
]
keywords = ["mcp", "powerpoint", "pptx", "s3", "aws"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-pptx>=1.0.2",
    "mcp>=1.1.2",
    "Pillow>=10.0.0",
    "fontTools>=4.42.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Atharvapaymode11/Office-PowerPoint-MCP-Server"
Repository = "https://github.com/Atharvapaymode11/Office-PowerPoint-MCP-Server"
Issues = "https://github.com/Atharvapaymode11/Office-PowerPoint-MCP-Server/issues"

[project.scripts]
ppt_mcp_server = "ppt_mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["tools", "utils"]
include = [
    "ppt_mcp_server.py",
    "slide_layout_templates.json",
]

[tool.hatch.build.targets.sdist]
include = [
    "/ppt_mcp_server.py",
    "/tools",
    "/utils",
    "/slide_layout_templates.json",
    "/README.md",
    "/pyproject.toml",
]